+++ docs/form_qc_coordinator/CHANGELOG.md
@@ -2,11 +2,14 @@
 
 All notable changes to this gear are documented in this file.
 
+## Unreleased
+* Updates to Python 3.12 and switches to use `fw-gear` instead of `flywheel-gear-toolkit` (now deprecated)
+
 ## 1.4.2
 * Updates to support legacy ingest pipeline
 * Evaluates subsequent visits only for longitudinal modules
 * Rebuilt for event logging updates
-  
+
 ## 1.4.1
 * Fixes a bug in updating the visit error log