+++ docs/redcap_project_info_management/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
+## Unreleased
+* Updates to Python 3.12 and switches to use `fw-gear` instead of `flywheel-gear-toolkit` (now deprecated)
+
 ## 1.0.0
 * Switch to major versioning
 * Rebuilt for ssm-parameter-store update